# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ALERYAF HUB API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/kpis:
    get:
      operationId: getDashboardKpis
      tags: [dashboard]
      summary: Get executive dashboard KPIs
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardKpis"
  /dashboard/daily-sales:
    get:
      operationId: getDailySales
      tags: [dashboard]
      summary: Get daily/monthly sales chart data
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
            enum: [TRY, USD]
        - name: branchId
          in: query
          schema:
            type: integer
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [daily, monthly]
      responses:
        "200":
          description: Sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySalesPoint"
  /branches:
    get:
      operationId: getBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchInput"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchInput"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /branches/analytics:
    get:
      operationId: getBranchAnalytics
      tags: [branches]
      summary: Get branch comparison analytics
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
            enum: [TRY, USD]
      responses:
        "200":
          description: Branch analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BranchAnalytics"
...
components:
  schemas:
...
    DashboardKpis:
      type: object
      properties:
        totalRevenueTry:
          type: number
        totalRevenueUsd:
          type: number
        totalCostTry:
          type: number
        totalCostUsd:
          type: number
        totalProfitTry:
          type: number
        totalProfitUsd:
          type: number
        totalInvoices:
          type: integer
        avgOrderValueTry:
          type: number
        avgOrderValueUsd:
          type: number
        inventoryValueTry:
          type: number
        inventoryValueUsd:
          type: number
        totalItems:
          type: integer
        lowStockCount:
          type: integer
        revenueGrowthTry:
          type: number
        revenueGrowthUsd:
          type: number
      required:
        - totalRevenueTry
        - totalRevenueUsd
        - totalCostTry
        - totalCostUsd
        - totalProfitTry
        - totalProfitUsd
        - totalInvoices
        - avgOrderValueTry
        - avgOrderValueUsd
        - inventoryValueTry
        - inventoryValueUsd
        - totalItems
        - lowStockCount
        - revenueGrowthTry
        - revenueGrowthUsd
    DailySalesPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        invoiceCount:
          type: integer
      required:
        - date
        - revenue
        - cost
        - profit
        - invoiceCount
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        code:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - isActive
        - createdAt
    CreateBranchInput:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        code:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - code
    BranchAnalytics:
      type: object
      properties:
        branchId:
          type: integer
        branchName:
          type: string
        branchCode:
          type: string
        revenueTry:
          type: number
        revenueUsd:
          type: number
        costTry:
          type: number
        costUsd:
          type: number
        profitTry:
          type: number
        profitUsd:
          type: number
        invoiceCount:
          type: integer
        contributionPctTry:
          type: number
        contributionPctUsd:
          type: number
      required:
        - branchId
        - branchName
        - branchCode
        - revenueTry
        - revenueUsd
        - costTry
        - costUsd
        - profitTry
        - profitUsd
        - invoiceCount
        - contributionPctTry
        - contributionPctUsd
...