# lib/api-zod/src/generated/types/createBranchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ALERYAF HUB API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBranchInput {
  name: string;
  nameAr?: string;
  code: string;
  isActive?: boolean;
}